pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "mammoth>=1.9.1",
    "opencv-python>=4.12.0.88",
    "openpyxl>=3.1.5",
    "pdf2docx>=0.5.8",
    "pdfplumber>=0.11.7",
    "pillow>=11.3.0",
    "pymupdf>=1.26.3",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "reportlab>=4.4.2",
]
